+++ projetotcc-back/src/main/java/br/edu/utfpr/pb/projetotcc/service/SubEtapaService.java
@@ -5,5 +5,5 @@ import br.edu.utfpr.pb.projetotcc.model.SubEtapa;
 import java.util.List;
 
 public interface SubEtapaService extends CrudService<SubEtapa, Long>{
-	List<SubEtapa> findAllByEtapaIdOrderByNome(Integer etapaId);
+	List<SubEtapa> findAllByEtapaIdOrderByNome(Long etapaId);
 }
